requirements.txt
absl-py==2.1.0
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.7.0
anyio==4.3.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
astroalign==2.5.1
astropy==6.0.1
astropy-iers-data==0.2024.4.1.0.33.14
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attrs==23.2.0
Babel==2.14.0
backoff==2.2.1
bcrypt==4.1.3
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.8.2
Brotli==1.1.0
bs4==0.0.2
build==1.2.1
cached-property==1.5.2
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb
ci-info==0.3.0
click==8.1.7
cloudpickle==3.0.0
cmasher==1.8.0
colorcet==3.1.0
coloredlogs==15.0.1
colorspacious==1.1.2
comm==0.2.2
configobj==5.0.8
configparser==7.0.0
contourpy==1.2.0
cycler==0.12.1
dataclasses-json==0.6.6
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
distro==1.9.0
dm-tree==0.1.8
dnspython==2.6.1
email_validator==2.1.1
entrypoints==0.4
etelemetry==0.3.1
exceptiongroup==1.2.0
executing==2.0.1
faiss-cpu==1.8.0
fastapi==0.111.0
fastapi-cli==0.0.4
fastjsonschema==2.19.1
filelock==3.13.3
flatbuffers==24.3.25
fonttools==4.50.0
fqdn==1.5.1
frontend==0.0.3
frozenlist==1.4.1
fsspec==2024.3.1
gast==0.5.4
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.29.0
google-pasta==0.2.0
googleapis-common-protos==1.63.0
greenlet==3.0.3
grpcio==1.62.1
h11==0.14.0
h2==4.1.0
h5py==3.10.0
hpack==4.0.0
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.1
humanfriendly==10.0
hyperframe==6.0.1
idna==3.6
imageio==2.34.0
importlib-metadata==7.0.0
importlib_resources==6.4.0
install==1.3.5
ipykernel==6.29.3
ipython==8.22.2
isodate==0.6.1
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
json5==0.9.24
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==7.4.9
jupyter_core==5.7.2
jupyter-events==0.10.0
jupyter-lsp==2.2.4
jupyter_server==2.13.0
jupyter_server_terminals==0.5.3
jupyterlab==4.1.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.4
keras==3.1.1
kiwisolver==1.4.5
kubernetes==29.0.0
langchain==0.2.1
langchain-community==0.2.1
langchain-core==0.2.1
langchain-text-splitters==0.2.0
langsmith==0.1.62
lazy_loader==0.4
libclang==18.1.1
llvmlite==0.42.0
looseversion==1.3.0
lxml==5.2.2
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.2
matplotlib==3.8.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.3.2
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
namex==0.0.7
nbclient==0.10.0
nbconvert==7.16.3
nbformat==5.10.4
nest_asyncio==1.6.0
networkx==3.2.1
nibabel==5.2.1
nipype==1.8.6
notebook==7.1.2
notebook_shim==0.2.4
numba==0.59.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.99
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
onnxruntime==1.18.0
openai==0.28.0
opencv-python==4.9.0.80
opentelemetry-api==1.24.0
opentelemetry-exporter-otlp-proto-common==1.24.0
opentelemetry-exporter-otlp-proto-grpc==1.24.0
opentelemetry-instrumentation==0.45b0
opentelemetry-instrumentation-asgi==0.45b0
opentelemetry-instrumentation-fastapi==0.45b0
opentelemetry-proto==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-semantic-conventions==0.45b0
opentelemetry-util-http==0.45b0
opt-einsum==3.3.0
optree==0.11.0
orjson==3.10.3
overrides==7.7.0
packaging==23.2
pandas==2.2.1
pandocfilters==1.5.0
parso==0.8.3
pathlib==1.0.1
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.2.0
pip==24.0
pkgutil_resolve_name==1.3.10
platformdirs==4.2.0
pooch==1.8.1
posthog==3.5.0
prometheus_client==0.20.0
prompt-toolkit==3.0.42
protobuf==4.25.3
prov==2.0.0
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==16.1.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
pydeck==0.9.1
pydot==2.0.0
pyerfa==2.0.1.1
Pygments==2.17.2
PyMuPDF==1.24.4
PyMuPDFb==1.24.3
pyparsing==3.1.2
pypdf==4.2.0
PyPika==0.48.9
pyproject_hooks==1.1.0
pyregion==2.2.0
PySocks==1.7.1
python-dateutil==2.9.0
python-dotenv==1.0.1
python-json-logger==2.0.7
python-multipart==0.0.9
pytz==2024.1
pyxnat==1.6.2
PyYAML==6.0.1
pysqlite3-binary==0.5.0
pyzmq==24.0.1
rdflib==7.0.0
referencing==0.34.0
regex==2024.5.15
regions==0.8
requests==2.31.0
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.18.0
rsa==4.9
safetensors==0.4.3
scikit-image==0.23.2
scikit-learn==1.4.1.post1
scipy==1.12.0
seaborn==0.13.2
Send2Trash==1.8.2
sentence-transformers==2.7.0
sep==1.2.1
setuptools==69.2.0
shellingham==1.5.4
simplejson==3.19.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
soxs==4.8.2
SQLAlchemy==2.0.30
stack-data==0.6.2
starlette==0.37.2
streamlit==1.35.0
sympy==1.12
tenacity==8.3.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-probability==0.24.0
termcolor==2.4.0
terminado==0.18.1
threadpoolctl==3.4.0
tifffile==2024.2.12
tiktoken==0.7.0
tinycss2==1.2.1
tokenizers==0.19.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.1
torch==2.2.2
torchvision==0.17.2
tornado==6.4
tqdm==4.66.2
traitlets==5.14.2
traits==6.3.2
transformers==4.41.1
typer==0.12.3
types-python-dateutil==2.9.0.20240316
typing_extensions==4.10.0
typing-inspect==0.9.0
typing-utils==0.1.0
tzdata==2024.1
ujson==5.10.0
uri-template==1.3.0
urllib3==2.2.1
uvicorn==0.29.0
uvloop==0.19.0
watchdog==4.0.1
watchfiles==0.21.0
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
websockets==12.0
Werkzeug==3.0.2
wheel==0.43.0
wrapt==1.16.0
yarl==1.9.4
zipp==3.17.0
